moving pre_tag functionality over to conf.py
