+++ openmdao/docs/Makefile
@@ -53,7 +53,6 @@ clean:
 
 html:
 	@rm -rf srcdocs
-	@python pre_tag.py
 	$(SPHINXBUILD) -b html $(ALLSPHINXOPTS) $(BUILDDIR)/html
 	@echo
 	@echo "Build finished. The HTML pages are in $(BUILDDIR)/html."
